+++ .changeset/pretty-beans-remain.md
@@ -1,0 +1,10 @@
+---
+'mocktomata': major
+'@mocktomata/framework': major
+---
+
+Fixing prototype chain.
+
+Remove `errorPlugin` as it is no longer needed.
+
+This change causes a breaking change as the `classPlugin` have changed it metadata.
