+++ packages/framework/package.json
@@ -72,6 +72,8 @@
 		"delay": "^5.0.0",
 		"depcheck": "^1.4.3",
 		"dirname-filename-esm": "^1.1.1",
+		"iso-error-google-cloud-api": "^5.2.0",
+		"iso-error-web": "^2.3.0",
 		"jest": "^29.4.2",
 		"jest-validate": "29.4.3",
 		"jest-watch-suspend": "^1.1.2",
