+++ packages/framework/ts/es2015.spec.ts
@@ -12,18 +12,11 @@ import {
 import { InvokeMetaMethodAfterSpec } from './spec/index.js'
 import {
 	callbackInDeepObjLiteral,
-	callbackInObjLiteral,
-	ChildOfDummy,
-	delayed,
-	Dummy,
-	fetch,
+	callbackInObjLiteral, delayed, fetch,
 	postReturn,
 	recursive,
 	simpleCallback,
-	synchronous,
-	WithProperty,
-	WithStaticMethod,
-	WithStaticProp
+	synchronous
 } from './test_artifacts/index.js'
 
 describe('basic checks', () => {
@@ -705,6 +698,16 @@ describe('function', () => {
 			await spec.done()
 		})
 	})
+
+	incubator('return object with number', (specName, spec) => {
+		test(specName, async () => {
+			const s = await spec(() => ({ a: 1 }))
+			const r = s()
+			expect(r).toEqual({ a: 1 })
+			await spec.done()
+		})
+	})
+
 	incubator('no input, array output', (specName, spec) => {
 		test(specName, async () => {
 			const subject = await spec(() => [1, 2, 'c'])
@@ -1163,653 +1166,6 @@ describe('promise', () => {
 	})
 })
 
-describe('class', () => {
-	class Foo {
-		constructor(public x: number) {}
-		getValue() {
-			return this.x
-		}
-	}
-
-	class Boo extends Foo {
-		getPlusOne() {
-			return this.getValue() + 1
-		}
-	}
-
-	incubator('invoke declared method', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Foo)
-			const instance = new Subject(1)
-			expect(instance.getValue()).toBe(1)
-			await spec.done()
-		})
-	})
-
-	incubator('invoke sub-class method', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Boo)
-
-			const instance = new Subject(1)
-			expect(instance.getPlusOne()).toBe(2)
-			await spec.done()
-		})
-	})
-
-	incubator('invoke parent method', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Boo)
-
-			const instance = new Subject(1)
-			expect(instance.getValue()).toBe(1)
-			await spec.done()
-		})
-	})
-
-	incubator('create multiple instances of the same class', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Foo)
-			const f1 = new Subject(1)
-			const f2 = new Subject(2)
-			expect(f1.getValue()).toBe(1)
-			expect(f2.getValue()).toBe(2)
-			await spec.done()
-		})
-	})
-
-	incubator.sequence('ok to use super/sub-class as long as behavior is the same', (specName, specs) => {
-		// It is ok to use diff
-		test(specName, async () => {
-			const save = specs.save
-			const bs = await save(Boo)
-			const boo = new bs(2)
-			expect(boo.getValue()).toBe(2)
-			await save.done()
-
-			const sim = specs.simulate
-			const fs = await sim(Foo)
-			const foo = new fs(2)
-			expect(foo.getValue()).toBe(2)
-			await sim.done()
-		})
-	})
-
-	class WithCallback {
-		callback(cb: (value: number) => void) {
-			setImmediate(() => {
-				cb(1)
-			})
-		}
-		justDo(x: any) {
-			return x
-		}
-	}
-	incubator('class method with callback', (specName, spec) => {
-		test(specName, async () => {
-			const s = await spec(WithCallback)
-			const cb = new s()
-
-			expect(cb.justDo(1)).toBe(1)
-			expect(
-				await new Promise(a => {
-					let total = 0
-					cb.callback(v => (total += v))
-					cb.callback(v => a(total + v))
-				})
-			).toBe(2)
-			await spec.done()
-		})
-	})
-
-	class Throwing {
-		doThrow() {
-			throw new Error('thrown')
-		}
-	}
-
-	incubator('invoke method throws', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Throwing)
-			const foo = new Subject()
-			a.throws(
-				() => foo.doThrow(),
-				e => e.message === 'thrown'
-			)
-			await spec.done()
-		})
-	})
-
-	class ResolvedPromise {
-		increment(x: number) {
-			return Promise.resolve(x + 1)
-		}
-	}
-
-	class DelayedPromise {
-		increment(x: number) {
-			return new Promise(a => {
-				setImmediate(() => a(x + 1))
-			})
-		}
-	}
-	incubator('method return resolved promise', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(ResolvedPromise)
-			const p = new Subject()
-			expect(await p.increment(3)).toBe(4)
-
-			await spec.done()
-		})
-	})
-
-	incubator('method returns delayed promise', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(DelayedPromise)
-			const p = new Subject()
-			expect(await p.increment(3)).toBe(4)
-
-			await spec.done()
-		})
-	})
-
-	incubator('invoke method returns delayed promise multiple times', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(DelayedPromise)
-			const p = new Subject()
-			expect(await Promise.all([p.increment(1), p.increment(3), p.increment(7)])).toEqual([2, 4, 8])
-
-			await spec.done()
-		})
-	})
-
-	class InvokeInternal {
-		do() {
-			return this.internal()
-		}
-		internal() {
-			return 'data'
-		}
-	}
-
-	incubator('method invokes internal method', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(InvokeInternal)
-			const a = new Subject()
-			expect(a.do()).toBe('data')
-
-			await spec.done()
-		})
-	})
-
-	class DelayedInvokeInternal {
-		getDelayedInner(delay = 0) {
-			return new Promise(a => {
-				setTimeout(() => {
-					a(this.inner())
-				}, delay)
-			})
-		}
-		inner() {
-			return 'inner'
-		}
-	}
-	incubator('method delay invokes internal method', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(DelayedInvokeInternal)
-			const a = new Subject()
-			expect(await a.getDelayedInner()).toBe('inner')
-			expect(a.inner()).toBe('inner')
-
-			await spec.done()
-		})
-	})
-
-	incubator.sequence('actual method is not invoked during simulation', (specName, { save, simulate }) => {
-		test(specName, async () => {
-			const Subject = await save(DelayedInvokeInternal)
-			const dii = new Subject()
-
-			expect(await dii.getDelayedInner()).toBe('inner')
-			await save.done()
-			{
-				class DelayedInvokeInternal {
-					getDelayedInner() {
-						throw new Error('should not call')
-					}
-				}
-				const BadSubject = await simulate(DelayedInvokeInternal)
-				const bad = new BadSubject()
-				expect(await bad.getDelayedInner()).toBe('inner')
-				await simulate.done()
-			}
-		})
-	})
-
-	class RejectLeak {
-		reject(x: number) {
-			return new Promise((_, r) => {
-				setImmediate(() => r(x))
-			})
-		}
-	}
-
-	incubator('runaway promise will not be leaked and break another test', (specName, spec) => {
-		test(`${specName}: setup`, async () => {
-			const MockRejector = await spec(RejectLeak)
-			const e = new MockRejector()
-			await a.throws(e.reject(300), v => v === 300)
-			await spec.done()
-		})
-		test(`${specName}: should not fail`, () => {
-			return new Promise<void>(a => setImmediate(() => a()))
-		})
-	})
-
-	class WithCircular {
-		value: any
-		cirRef: WithCircular
-		constructor() {
-			this.cirRef = this
-		}
-	}
-
-	class ClassWithCircular {
-		channel: WithCircular
-		constructor() {
-			this.channel = new WithCircular()
-		}
-		exec(cmd: string, cb: AnyFunction) {
-			this.channel.value = cmd
-			cb(this.channel)
-		}
-	}
-
-	incubator('can use class with circular reference', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(ClassWithCircular)
-			const f = new Subject()
-
-			let actual
-			f.exec('echo', (channel: any) => {
-				actual = channel.value
-			})
-
-			expect(actual).toBe('echo')
-			await spec.done()
-		})
-	})
-
-	incubator('class with circular reference accessing', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(ClassWithCircular)
-			const f = new Subject()
-
-			let actual
-			f.exec('echo', (channel: WithCircular) => {
-				actual = channel.cirRef.value
-			})
-
-			expect(actual).toBe('echo')
-			await spec.done()
-		})
-	})
-
-	class Channel {
-		listeners: any[] = []
-		stdio: any
-		constructor() {
-			this.stdio = this
-		}
-		on(listener: any) {
-			this.listeners.push(listener)
-		}
-		emit(data: any) {
-			this.listeners.forEach(l => l(data))
-		}
-	}
-	class Ssh {
-		channel: Channel
-		constructor() {
-			this.channel = new Channel()
-		}
-		exec(cmd: string, cb: AnyFunction) {
-			cb(this.channel)
-			this.channel.stdio.emit(cmd)
-		}
-	}
-
-	// TODO: throws NotSupported error for callback with complex object.
-	// This gives indication to the user that a plugin is need to support this subject
-	// To fix this, I need to:
-	// 1. get property key and value from object without invoking getter.
-	// 2. Add GetAction SetAction back
-	incubator('callback with complex object', (specName, spec) => {
-		test(specName, async () => {
-			const Subject = await spec(Ssh)
-			const f = new Subject()
-
-			let actual
-			f.exec('echo', (channel: any) => channel.stdio.on((data: any) => (actual = data)))
-
-			expect(actual).toBe('echo')
-			await spec.done()
-		})
-	})
-
-	incubator('use composite callback function', (specName, spec) => {
-		test(specName, async () => {
-			class Foo {
-				on(compositeFn: any) {
-					return this.internal(compositeFn)
-				}
-				internal(input: any) {
-					expect(input.value).toBe('xyz')
-					return input
-				}
-			}
-			const fn = Object.assign(
-				function () {
-					return
-				},
-				{
-					value: 'xyz'
-				}
-			)
-			const s = await spec(Foo)
-			const f = new s()
-			const actual = f.on(fn)
-
-			expect(actual.value).toBe('xyz')
-
-			await spec.done()
-		})
-	})
-
-	incubator('class with property', (specName, spec) => {
-		test(specName, async () => {
-			const s = await spec(WithProperty)
-			const p = new s()
-			expect(p.do(2)).toBe(2)
-			expect(p.y).toBe(1)
-			p.y = 3
-			expect(p.y).toBe(3)
-			await spec.done()
-		})
-	})
-
-	incubator('static property', (specName, spec) => {
-		test(specName, async () => {
-			const s = await spec(WithStaticProp)
-			expect(s.x).toBe(1)
-			expect((s.x = 3)).toBe(3)
-			await spec.done()
-		})
-	})
-
-	incubator('static method', (specName, spec) => {
-		test(specName, async () => {
-			const s = await spec(WithStaticMethod)
-			expect(s.do()).toBe('foo')
-			await spec.done()
-		})
-	})
-})
-
-describe('instance', () => {
-	incubator('passes instanceof test for 1st level class', (specName, spec) => {
-		test(specName, async () => {
-			const S = await spec(Dummy)
-			const s = new S()
-			expect(s).toBeInstanceOf(Dummy)
-			await spec.done()
-		})
-	})
-
-	incubator('passes instanceof test for sub class', (specName, spec) => {
-		test(specName, async () => {
-			const S = await spec(ChildOfDummy)
-			const s = new S()
-			expect(s).toBeInstanceOf(ChildOfDummy)
-			expect(s).toBeInstanceOf(Dummy)
-			await spec.done()
-		})
-	})
-
-	incubator(
-		'instanceof for output instance is not supported (need custom plugin for this)',
-		(specName, spec) => {
-			test.skip(specName, async () => {
-				function fool() {
-					return new ChildOfDummy()
-				}
-				const f = await spec(fool)
-				const s = f()
-				expect(s).toBeInstanceOf(ChildOfDummy)
-				expect(s).toBeInstanceOf(Dummy)
-				await spec.done()
-			})
-		}
-	)
-
-	incubator('instanceof for output instance if the class was used through input', (specName, spec) => {
-		test(specName, async () => {
-			const subject = {
-				in(_: any) {
-					return
-				},
-				out() {
-					return new ChildOfDummy()
-				}
-			}
-			const s = await spec(subject)
-			s.in(new ChildOfDummy())
-			const actual = s.out()
-			expect(actual).toBeInstanceOf(ChildOfDummy)
-			expect(actual).toBeInstanceOf(Dummy)
-			await spec.done()
-		})
-	})
-
-	incubator.sequence('getter skips internal method calls', (specName, { save, simulate }) => {
-		test(specName, async () => {
-			const spy = await save({
-				get side() {
-					this.internalCall()
-					return 'ok'
-				},
-				internalCall() {}
-			})
-			expect(spy.side).toBe('ok')
-			await save.done()
-
-			const stub = await simulate({
-				get side() {
-					this.internalCall()
-					return 'ok'
-				},
-				internalCall() {
-					throw new Error('should not reach')
-				}
-			})
-			expect(stub.side).toBe('ok')
-			await simulate.done()
-		})
-	})
-
-	incubator.sequence('setter skips internal method calls', (specName, { save, simulate }) => {
-		test(specName, async () => {
-			const spy = await save({
-				set side(v: any) {
-					this.internalCall()
-				},
-				internalCall() {}
-			})
-			spy.side = 1
-			await save.done()
-
-			const stub = await simulate({
-				set side(v: any) {
-					this.internalCall()
-				},
-				internalCall() {
-					throw new Error('should not reach')
-				}
-			})
-			stub.side = 1
-			await simulate.done()
-		})
-	})
-
-	incubator.sequence('method skips internal method calls', (specName, { save, simulate }) => {
-		test(specName, async () => {
-			let shouldThrow = false
-			class InvokeInternal {
-				foo() {
-					this.internalCall()
-					return 'ok'
-				}
-				internalCall() {
-					if (shouldThrow) throw new Error('should not reach')
-				}
-			}
-			const instance = new InvokeInternal()
-			const spy = await save(instance)
-			expect(spy.foo()).toBe('ok')
-			await save.done()
-			shouldThrow = true
-
-			const stub = await simulate(instance)
-			expect(stub.foo()).toBe('ok')
-			await simulate.done()
-		})
-	})
-
-	incubator('returning this is same as spy', (specName, spec) => {
-		test(specName, async () => {
-			class Fluent {
-				foo() {
-					return this
-				}
-			}
-
-			const spy = await spec(new Fluent())
-			expect(spy.foo()).toBe(spy)
-			await spec.done()
-		})
-	})
-
-	incubator('static property', (specName, spec) => {
-		test(specName, async () => {
-			WithStaticProp.x = 1
-			function getClass() {
-				return WithStaticProp
-			}
-			const subject = await spec(getClass)
-			const s = subject()
-			expect(s.x).toBe(1)
-			expect((s.x = 3)).toBe(3)
-			await spec.done()
-		})
-	})
-
-	incubator('static method', (specName, spec) => {
-		test(specName, async () => {
-			function getClass() {
-				return WithStaticMethod
-			}
-
-			const subject = await spec(getClass)
-			const s = subject()
-			expect(s.do()).toBe('foo')
-			await spec.done()
-		})
-	})
-
-	incubator('constructor throws error', (specName, spec) => {
-		test(specName, async () => {
-			class Throw {
-				constructor(x: string) {
-					throw new Error(x)
-				}
-				// need this dummy method for the system to detect this is a class
-				foo() {}
-			}
-			const s = await spec({ Throw })
-			const err = a.throws(() => new s.Throw('abc'))
-
-			expect(err.message).toBe('abc')
-
-			await spec.done()
-		})
-	})
-
-	incubator.sequence('instantiate with wrong primitive argument', (specName, { save, simulate }) => {
-		test(specName, async () => {
-			class EchoConstructorArg {
-				constructor(public value: number) {}
-				echo() {
-					return this.value
-				}
-			}
-
-			const s = await save({ EchoConstructorArg })
-			new s.EchoConstructorArg(1)
-			await save.done()
-
-			const s2 = await simulate({ EchoConstructorArg })
-			a.throws(() => new s2.EchoConstructorArg(2), ActionMismatch)
-		})
-	})
-
-	incubator.sequence(
-		'instantiate with different argument is okay as long as behavior does not change',
-		(specName, { save, simulate }) => {
-			test(specName, async () => {
-				class EchoConstructorArg {
-					constructor(public value: string) {}
-					echo() {
-						return this.value
-					}
-				}
-
-				const s = await save({ EchoConstructorArg })
-				new s.EchoConstructorArg('abc')
-				await save.done()
-
-				const s2 = await simulate({ EchoConstructorArg })
-				new s2.EchoConstructorArg('xyz')
-				await simulate.done()
-			})
-		}
-	)
-
-	incubator('ioc instantiate class', (specName, spec) => {
-		test(specName, async () => {
-			class Dummy {
-				foo() {}
-			}
-			const s = await spec((subject: any) => new subject())
-			expect(s(Dummy)).toBeInstanceOf(Dummy)
-
-			await spec.done()
-		})
-	})
-})
-
-describe('error', () => {
-	incubator('set error property', (specName, spec) => {
-		it(specName, async () => {
-			const s = await spec(() => {
-				throw new Error('ha')
-			})
-			const err: any = a.throws(s)
-			err.x = 1
-			expect(err.x).toEqual(1)
-			await spec.done()
-		})
-	})
-})
-
 describe('ignoreMismatch', () => {
 	incubator.save('call after spec throws', (specName, spec) => {
 		test(specName, async () => {
@@ -1939,24 +1295,6 @@ describe('symbol', () => {
 		it(specName, async () => {
 			const s = await spec(takeSymbol)
 			expect(s(Symbol.for('abc'))).toEqual(Symbol.for('abc'))
-			await spec.done()
-		})
-	})
-})
-
-describe('generator', () => {
-	function* foo() {
-		yield 'a'
-		yield 'b'
-		yield 'c'
-	}
-	incubator('yield', (specName, spec) => {
-		it(specName, async () => {
-			const s = await spec(foo)
-			const i = s()
-			expect(i.next()).toEqual({ done: false, value: 'a' })
-			expect(i.next()).toEqual({ done: false, value: 'b' })
-			expect(i.next()).toEqual({ done: false, value: 'c' })
 			await spec.done()
 		})
 	})
