+++ packages/framework/ts/index.ts
@@ -11,7 +11,6 @@ export * from './mockto/index.js'
 export * from './plugins/array/index.js'
 export * from './plugins/bigint/index.js'
 export * from './plugins/class/index.js'
-export * from './plugins/error/index.js'
 export * from './plugins/function/index.js'
 export * from './plugins/inertValue/index.js'
 export * from './plugins/instance/index.js'
