--- packages/framework/ts/plugins/error/index.ts
@@ -1,1 +1,0 @@
-export * from './error_plugin.js'
