+++ packages/framework/ts/plugins/instance/instance_plugin.ts
@@ -46,7 +46,8 @@ export const instancePlugin: SpecPlugin<
 		const classProto = Object.getPrototypeOf(classConstructor)
 		// `classProto` can be `{}` which has no prototype.
 		// default to `null` in that case
-		Object.setPrototypeOf(base, classProto.prototype ?? null)
+		const proto = classProto.prototype ? Object.create(classProto.prototype) : null
+		Object.setPrototypeOf(base, proto)
 		const stub = new Proxy(base, {
 			get(_: any, property: string) {
 				if (meta.functionCalls.length > 0 && meta.functionCalls[0] === property) {
