+++ packages/framework/ts/testutils/index.ts
@@ -1,3 +1,4 @@
+export * from './caller.js'
+export * from './proto.js'
 export * from './test_context.js'
 export * from './test_io.js'
-export * from './caller.js'
