+++ tests/nodejs/.mocktomata/specs/ts/axios.accept.ts/1-+-1-=-2
@@ -1,7 +1,7 @@
 1 + 1 = 2
 {
   "refs": [
-    {"plugin":"@mocktomata/function","profile":"target","meta":"{\"type\":\"function\",\"name\":\"wrap\",\"length\":0,\"props\":{\"VERSION\":\"1.1.3\"}}"},
+    {"plugin":"@mocktomata/function","profile":"target","meta":"{\"type\":\"function\",\"name\":\"wrap\",\"length\":0,\"props\":{\"VERSION\":\"1.3.3\"}}"},
     {"plugin":"@mocktomata/function","profile":"target","source":{"type":"result","id":0},"meta":"{\"type\":\"function\",\"name\":\"wrap\",\"length\":0,\"props\":{}}"},
     {"plugin":"@mocktomata/string","profile":"input","source":{"type":"argument","id":2,"key":0},"meta":"http://api.mathjs.org/v4/?expr=1%2B1"},
     {"plugin":"@mocktomata/promise","profile":"output","source":{"type":"result","id":2},"meta":"resolved-promise"},
@@ -12,11 +12,11 @@
   "actions": [
     {"type":"get","refId":"0","performer":"user","tick":0,"key":"get"},
     {"type":"return","actionId":0,"tick":0,"payload":"1"},
-    {"type":"invoke","refId":"1","performer":"user","thisArg":"0","payload":["2"],"tick":1},
-    {"type":"return","actionId":2,"tick":16,"payload":"3"},
-    {"type":"invoke","refId":"3","performer":"plugin","site":"then","thisArg":"4","payload":["5"],"tick":218},
-    {"type":"invoke","refId":"5","performer":"mockto","thisArg":"4","payload":["6"],"tick":0},
-    {"type":"return","actionId":5,"tick":1,"payload":"4"},
+    {"type":"invoke","refId":"1","performer":"user","thisArg":"0","payload":["2"],"tick":0},
+    {"type":"return","actionId":2,"tick":20,"payload":"3"},
+    {"type":"invoke","refId":"3","performer":"plugin","site":"then","thisArg":"4","payload":["5"],"tick":212},
+    {"type":"invoke","refId":"5","performer":"mockto","thisArg":"4","payload":["6"],"tick":1},
+    {"type":"return","actionId":5,"tick":0,"payload":"4"},
     {"type":"return","actionId":4,"tick":0,"payload":"4"},
     {"type":"get","refId":"6","performer":"user","tick":0,"key":"data"},
     {"type":"return","actionId":8,"tick":0,"payload":2}
