+++ tests/nodejs/.mocktomata/specs/ts/axios.zucchini.accept.ts/axios-with-error
@@ -1,12 +1,14 @@
 axios with error
 {
   "refs": [
-    {"plugin":"@mocktomata/function","profile":"target","meta":"{\"type\":\"function\",\"name\":\"wrap\",\"length\":0,\"props\":{\"VERSION\":\"1.1.3\"}}"},
+    {"plugin":"@mocktomata/function","profile":"target","meta":"{\"type\":\"function\",\"name\":\"wrap\",\"length\":0,\"props\":{\"VERSION\":\"1.3.3\"}}"},
     {"plugin":"@mocktomata/undefined","profile":"input","source":{"type":"this","id":0}},
     {"plugin":"@mocktomata/string","profile":"input","source":{"type":"argument","id":0,"key":0},"meta":"http://api.mathjs.org/v4/?expr=1+1"},
     {"plugin":"@mocktomata/promise","profile":"output","source":{"type":"result","id":0},"meta":"rejected-promise"},
     {"plugin":"@mocktomata/function","profile":"input","source":{"type":"argument","id":2,"key":1},"meta":"{\"type\":\"function\",\"name\":\"\",\"length\":1,\"props\":{}}"},
-    {"plugin":"@mocktomata/error","profile":"output","source":{"type":"argument","id":3,"key":0},"meta":{"err":"{\"type\":\"object\",\"props\":{\"message\":\"Request failed with status code 400\",\"name\":\"AxiosError\",\"code\":\"ERR_BAD_REQUEST\"}}","functionCalls":[]}},
+    {"plugin":"@mocktomata/instance","profile":"output","source":{"type":"argument","id":3,"key":0},"meta":{"base":"{\"type\":\"object\",\"props\":{\"message\":\"Request failed with status code 400\",\"name\":\"AxiosError\",\"code\":\"ERR_BAD_REQUEST\"}}","classConstructor":"6","functionCalls":[]}},
+    {"plugin":"@mocktomata/class","profile":"output","meta":{"meta":"{\"type\":\"function\",\"name\":\"AxiosError\",\"length\":5,\"props\":{}}","parent":"7"}},
+    {"plugin":"@mocktomata/function","profile":"output","meta":"{\"type\":\"function\",\"name\":\"\",\"length\":0,\"props\":{}}"},
     {"plugin":"@mocktomata/string","profile":"output","source":{"type":"result","id":6},"meta":"AxiosError"},
     {"plugin":"@mocktomata/string","profile":"output","source":{"type":"result","id":8},"meta":"ERR_BAD_REQUEST"},
     {"plugin":"@mocktomata/object","profile":"output","source":{"type":"result","id":10},"meta":"{\"type\":\"object\",\"props\":{\"status\":400,\"statusText\":\"Bad Request\",\"data\":\"Error: Unexpected part \\\"1\\\" (char 3)\"}}"},
@@ -16,25 +18,25 @@ axios with error
   "actions": [
     {"type":"invoke","refId":"0","performer":"user","thisArg":"1","payload":["2"],"tick":0},
     {"type":"return","actionId":0,"tick":1,"payload":"3"},
-    {"type":"invoke","refId":"3","performer":"plugin","site":"then","thisArg":"1","payload":["1","4"],"tick":171},
+    {"type":"invoke","refId":"3","performer":"plugin","site":"then","thisArg":"1","payload":["1","4"],"tick":175},
     {"type":"invoke","refId":"4","performer":"mockto","thisArg":"1","payload":["5"],"tick":0},
-    {"type":"return","actionId":3,"tick":2,"payload":"1"},
+    {"type":"return","actionId":3,"tick":1,"payload":"1"},
     {"type":"return","actionId":2,"tick":0,"payload":"1"},
     {"type":"get","refId":"5","performer":"user","tick":0,"key":"name"},
-    {"type":"return","actionId":6,"tick":0,"payload":"6"},
+    {"type":"return","actionId":6,"tick":0,"payload":"8"},
     {"type":"get","refId":"5","performer":"user","tick":0,"key":"code"},
-    {"type":"return","actionId":8,"tick":0,"payload":"7"},
+    {"type":"return","actionId":8,"tick":0,"payload":"9"},
     {"type":"get","refId":"5","performer":"user","tick":0,"key":"response"},
-    {"type":"return","actionId":10,"tick":0,"payload":"8"},
-    {"type":"get","refId":"8","performer":"user","tick":0,"key":"status"},
+    {"type":"return","actionId":10,"tick":0,"payload":"10"},
+    {"type":"get","refId":"10","performer":"user","tick":0,"key":"status"},
     {"type":"return","actionId":12,"tick":0,"payload":400},
-    {"type":"get","refId":"5","performer":"user","tick":1,"key":"response"},
-    {"type":"return","actionId":14,"tick":0,"payload":"8"},
-    {"type":"get","refId":"8","performer":"user","tick":0,"key":"statusText"},
-    {"type":"return","actionId":16,"tick":0,"payload":"9"},
     {"type":"get","refId":"5","performer":"user","tick":0,"key":"response"},
-    {"type":"return","actionId":18,"tick":0,"payload":"8"},
-    {"type":"get","refId":"8","performer":"user","tick":0,"key":"data"},
-    {"type":"return","actionId":20,"tick":0,"payload":"10"}
+    {"type":"return","actionId":14,"tick":0,"payload":"10"},
+    {"type":"get","refId":"10","performer":"user","tick":0,"key":"statusText"},
+    {"type":"return","actionId":16,"tick":0,"payload":"11"},
+    {"type":"get","refId":"5","performer":"user","tick":0,"key":"response"},
+    {"type":"return","actionId":18,"tick":0,"payload":"10"},
+    {"type":"get","refId":"10","performer":"user","tick":0,"key":"data"},
+    {"type":"return","actionId":20,"tick":0,"payload":"12"}
   ]
 }
