+++ tests/nodejs/package.json
@@ -41,6 +41,7 @@
 		"@types/ws": "^8.5.4",
 		"assertron": "^11.0.0",
 		"axios": "^1.1.3",
+		"axios-mock-adapter": "^1.21.2",
 		"cross-env": "^7.0.3",
 		"cross-fetch": "^3.1.4",
 		"depcheck": "^1.4.3",
@@ -60,5 +61,10 @@
 	},
 	"engines": {
 		"node": ">= 14.16"
+	},
+	"dependencies": {
+		"iso-error": "^6.0.0",
+		"iso-error-google-cloud-api": "^5.2.0",
+		"iso-error-web": "^2.3.0"
 	}
 }
